backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
pydantic
python-dotenv
pytest
flask-migrate
